pyproject.toml
[tool.poetry]
name = "photobackup"
version = "0.1.0"
description = ""
authors = ["Pawel Zukowski <p.z.idlecode@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.1.2"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
